writing KML directly to file instead of abusing use of StringBuilder + minor improvements
I overloaded writeKMLFile() in KmlWriter class to be a boolean and to wirte directly to file.
I still keep old code as it is pretty handy for outputing text and being able to debug strings.

I had to overload writeAllDataRecords() in a not very DRY way to avoid issues with changing declared type of variable. I would prefer to have it nice and DRY but it will have to do for some time. Will hopefully be deleted later.

setMessageLabelText() can be used for non-errors now.